requirements.txt
torch>=2.2.0
torchvision>=0.17.0
timm>=0.9.12
opencv-python>=4.8.0
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.13
pillow>=10.0
tqdm>=4.66
streamlit>=1.30
grad-cam>=1.4.8
python-dotenv>=1.0
